regulator: act8846: fix incorrect gpio probing
The gpios are incorrectly assigned, making the error code being
assigned to the gpio number when an entry is not found in the DT.
The first visible side effect is that act8846 fails like this :

[    0.740895] act8846_i2c_probe,line=898
[    0.742312] act8846_parse_dt,line=746
[    0.743641] of_get_named_gpio_flags exited with status 225
[    0.743689] of_get_named_gpio_flags exited with status 226
[    0.743727] of_get_named_gpio_flags: can't parse gpios property
[    0.743758] invalid gpio: -2
[    0.743818] of_get_named_gpio_flags: can't parse gpios property
[    0.743848] lhm invalid gpio: -2
[    0.743937] act8846_i2c_probe: act8846_pmic_hold=1
[    0.744034] act8846_i2c_probe: act8846_pmic_sleep=1
[    0.744101] ------------[ cut here ]------------
[    0.744199] WARNING: at drivers/gpio/gpiolib.c:126 gpio_to_desc+0x20/0x44()
[    0.744278] invalid GPIO -2
[    0.744330] Modules linked in:
[    0.744410] CPU: 0 PID: 1 Comm: swapper/0 Not tainted 3.10.101 #6
[    0.744534] [<c00143b0>] (unwind_backtrace+0x0/0xe0) from [<c0011894>] (show_stack+0x10/0x14)
[    0.744669] [<c0011894>] (show_stack+0x10/0x14) from [<c002dfd4>] (warn_slowpath_common+0x4c/0x6c)
(...)

The second side effect is that it disables vdd_logic and forces
the dvfs driver to remain at the lower possible voltage for the
DDR, which is 200 MHz, as noticed here :

[    2.109611] ddrfreq: normal 396MHz video_1080p 0MHz video_4k 396MHz dualview 0MHz idle 0MHz suspend 200MHz reboot 396MHz
...
[    2.141172] ddrfreq: change freq to 200 MHz when normal

This patch provides a simple fix for the incorrect values. Please note
that a few other places continue to reference them and that it should
be cleaner to only assign the results when it's known they're valid. At
least this fix allows the system to work properly now.