rockchip: dvfs: add default min and max scaling frequencies in the DT
When trying to adapt to various boards whose heat sinking capabilities
vary, it's convenient to be able to run with reliable default limits
and let the user unlock them.

This patch adds default-min-freq and default-max-freq to the device
tree, in clk_core next to pvtm-opertaing-points. When they are set,
the values are applied as the initial scaling_min_freq and
scaling_max_freq values.

This way it is possible to specify extra frequencies for boards equipped
with larger heatsinks.